+++ packages/hightide-design/src/component-token-resolvers/icon-button-tokens.ts
@@ -123,6 +123,14 @@ export const iconButtonTokenResolver: IconButtonTokenResolver = ({
     },
     stateLayer: {
       backgroundColor: tint,
+      position: {
+        type: 'absolute',
+        top: 0,
+        right: 0,
+        bottom: 0,
+        left: 0,
+        zIndex: 20,
+      },
       shape: {
         borderRadius: { type: 'all', value: layout.borderRadius },
       },