+++ packages/hightide-design/src/component-token-resolvers/pressable-tokens.ts
@@ -154,6 +154,14 @@ export const pressableTokenResolver: PressableTokenResolver = ({
     },
     stateLayer: {
       backgroundColor: tint,
+      position: {
+        type: 'absolute',
+        top: 0,
+        right: 0,
+        bottom: 0,
+        left: 0,
+        zIndex: 20,
+      },
       shape: {
         borderRadius: { type: 'all', value: layout.borderRadius },
       },