+++ packages/hightide-native/src/components/user-interaction/Textarea.tsx
@@ -3,10 +3,9 @@ import {
   useMemo,
   useState
 } from 'react'
-import type { TextStyle } from 'react-native'
 import {
-  StyleSheet,
   TextInput,
+  View,
   type TextInputProps
 } from 'react-native'
 
@@ -21,6 +20,7 @@ import { useMemoizedTheme } from '../../hooks/useMemoizedTheme'
 import type {
   TextareaContainerStyle,
   TextareaState,
+  TextareaStateLayerStyle,
   TextareaTextStyle
 } from '../../theme/types/components/textarea'
 import type { StyleOverwrite } from '../../theme/types/resolver'
@@ -45,6 +45,7 @@ export type TextareaProps = Omit<TextInputProps, 'value' | 'style'>
     editCompleteOptions?: EditCompleteOptions,
     initialValue?: string,
     style?: StyleOverwrite<TextareaState, TextareaContainerStyle>,
+    stateLayerStyle?: StyleOverwrite<TextareaState, TextareaStateLayerStyle>,
     textStyle?: StyleOverwrite<TextareaState, TextareaTextStyle>,
   }
 
@@ -60,7 +61,10 @@ export const Textarea = forwardRef<TextInput, TextareaProps>(function Textarea({
   onEditComplete,
   editCompleteOptions,
   style,
+  stateLayerStyle,
   textStyle,
+  onPointerEnter,
+  onPointerLeave,
   ...props
 }, ref) {
   const { theme } = useTheme()
@@ -99,80 +103,77 @@ export const Textarea = forwardRef<TextInput, TextareaProps>(function Textarea({
   }), [color, disabled, invalid, readOnly, interactive, isHovered, isPressed, isFocused])
 
   const resolvedContainerStyle = useMemoizedTheme(theme.components.textarea.container, state, style)
+  const resolvedStateLayerStyle = useMemoizedTheme(theme.components.textarea.stateLayer, state, stateLayerStyle)
   const resolvedTextStyle = useMemoizedTheme(theme.components.textarea.text, state, textStyle)
   const resolvedPlaceholderStyle = useMemoizedTheme(theme.components.textarea.placeholder, state)
 
-  const resolvedStyle = useMemo(
-    () => {
-      const resolvedSheet: TextStyle = StyleSheet.flatten<TextStyle>([
-        resolvedContainerStyle,
-        resolvedTextStyle,
-      ])
-      delete resolvedSheet['boxShadow']
-      return resolvedSheet
-    },
-    [resolvedContainerStyle, resolvedTextStyle]
-  )
-
   return (
-    <TextInput
-      {...props}
-      ref={ref}
-      value={value}
-      editable={interactive}
-      multiline
-      textAlignVertical="top"
-      placeholderTextColor={resolvedPlaceholderStyle.color}
-      onChangeText={(nextValue) => {
-        props.onChangeText?.(nextValue)
-        restartTimer(() => {
-          onEditComplete?.(nextValue)
-        })
-        setValue(nextValue)
-      }}
-      onFocus={(event) => {
-        if (interactive) {
-          setIsFocused(true)
-        }
-        props.onFocus?.(event)
-      }}
-      onBlur={(event) => {
-        setIsFocused(false)
-        props.onBlur?.(event)
-        if (allowEditCompleteOnBlur) {
-          onEditComplete?.(value ?? '')
-          clearTimer()
-        }
-      }}
-      onPressIn={(event) => {
-        if (interactive) {
-          setIsPressed(true)
-        }
-        props.onPressIn?.(event)
-      }}
-      onPressOut={(event) => {
-        setIsPressed(false)
-        props.onPressOut?.(event)
-      }}
+    <View
+      style={resolvedContainerStyle}
       onPointerEnter={(event) => {
         if (interactive) {
           setIsHovered(true)
         }
-        props.onPointerEnter?.(event)
+        onPointerEnter?.(event)
       }}
       onPointerLeave={(event) => {
         setIsHovered(false)
-        props.onPointerLeave?.(event)
+        onPointerLeave?.(event)
       }}
-      onSubmitEditing={(event) => {
-        props.onSubmitEditing?.(event)
-        if (allowEnterComplete) {
-          onEditComplete?.(value ?? '')
-          clearTimer()
-        }
-      }}
-      style={resolvedStyle}
-      accessibilityState={{ disabled, selected: required }}
-    />
+    >
+      <View
+        pointerEvents="none"
+        style={resolvedStateLayerStyle}
+      />
+      <TextInput
+        {...props}
+        ref={ref}
+        value={value}
+        editable={interactive}
+        multiline
+        textAlignVertical="top"
+        placeholderTextColor={resolvedPlaceholderStyle.color}
+        onChangeText={(nextValue) => {
+          props.onChangeText?.(nextValue)
+          restartTimer(() => {
+            onEditComplete?.(nextValue)
+          })
+          setValue(nextValue)
+        }}
+        onFocus={(event) => {
+          if (interactive) {
+            setIsFocused(true)
+          }
+          props.onFocus?.(event)
+        }}
+        onBlur={(event) => {
+          setIsFocused(false)
+          props.onBlur?.(event)
+          if (allowEditCompleteOnBlur) {
+            onEditComplete?.(value ?? '')
+            clearTimer()
+          }
+        }}
+        onPressIn={(event) => {
+          if (interactive) {
+            setIsPressed(true)
+          }
+          props.onPressIn?.(event)
+        }}
+        onPressOut={(event) => {
+          setIsPressed(false)
+          props.onPressOut?.(event)
+        }}
+        onSubmitEditing={(event) => {
+          props.onSubmitEditing?.(event)
+          if (allowEnterComplete) {
+            onEditComplete?.(value ?? '')
+            clearTimer()
+          }
+        }}
+        style={resolvedTextStyle}
+        accessibilityState={{ disabled, selected: required }}
+      />
+    </View>
   )
 })